Refactor item list and Add active tab color
